requirements.txt
ocrmypdf
whisper
Pillow
transformers
sentence-transformers
faiss-cpu
peft
langchain
crewai
langgraph
fastapi
uvicorn
